@@ -1,23 +1,23 @@
-import { Resolve, Router, ActivatedRouteSnapshot } from '@angular/router';
-import { User } from '../_models/user';
-import { Injectable } from '@angular/core';
+import {Injectable} from '@angular/core';
+import {User} from '../_models/user';
+import {Resolve, Router, ActivatedRouteSnapshot} from '@angular/router';
 import { UserService } from '../_services/user.service';
 import { AlertifyService } from '../_services/alertify.service';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/of';
-import 'rxjs/add/operator/catch';
+import { Observable, of } from 'rxjs';
+import { catchError } from 'rxjs/operators';
 
 @Injectable()
 export class MemberDetailResolver implements Resolve<User> {
-
-    constructor(private userService: UserService,
-        private router: Router, private alertify: AlertifyService) {}
+    constructor(private userService: UserService, private router: Router,
+        private alertify: AlertifyService) {}
 
     resolve(route: ActivatedRouteSnapshot): Observable<User> {
-        return this.userService.getUser(route.params['id']).catch(error => {
-            this.alertify.error('Problem retrieving data');
-            this.router.navigate(['/members']);
-            return Observable.of(null);
-        });
+        return this.userService.getUser(route.params['id']).pipe(
+            catchError(error => {
+                this.alertify.error('Problem retrieving data');
+                this.router.navigate(['/members']);
+                return of(null);
+            })
+        );
     }
 }
