@@ -1,23 +1,24 @@
-import { Resolve, Router, ActivatedRouteSnapshot } from '@angular/router';
-import { User } from '../_models/user';
-import { Injectable } from '@angular/core';
+import {Injectable} from '@angular/core';
+import {User} from '../_models/user';
+import {Resolve, Router, ActivatedRouteSnapshot} from '@angular/router';
 import { UserService } from '../_services/user.service';
 import { AlertifyService } from '../_services/alertify.service';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/of';
-import 'rxjs/add/operator/catch';
+import { Observable } from 'rxjs';
+import { catchError } from 'rxjs/operators';
+import { of } from 'rxjs';
 
 @Injectable()
 export class MemberListResolver implements Resolve<User[]> {
-
-    constructor(private userService: UserService,
-        private router: Router, private alertify: AlertifyService) {}
+    constructor(private userService: UserService, private router: Router,
+        private alertify: AlertifyService) {}
 
     resolve(route: ActivatedRouteSnapshot): Observable<User[]> {
-        return this.userService.getUsers().catch(error => {
-            this.alertify.error('Problem retrieving data');
-            this.router.navigate(['/members']);
-            return Observable.of(null);
-        });
+        return this.userService.getUsers().pipe(
+            catchError(error => {
+                this.alertify.error('Problem retrieving data');
+                this.router.navigate(['/home']);
+                return of(null);
+            })
+        );
     }
 }
