@@ -117,8 +117,9 @@ namespace DatingApp.API.Controllers
                 return BadRequest("This is photo is main already");
 
             var mainActualPhoto = await _repo.GetUserMainPhoto(userId);
-            mainActualPhoto.IsMain = false;
-
+            
+                mainActualPhoto.IsMain = false;
+            
             photoFromRepo.IsMain = true;
 
             if(await _repo.SaveAll())
@@ -129,6 +130,46 @@ namespace DatingApp.API.Controllers
 
         }
         
-        
+           [HttpDelete("{id}")]
+        public async Task<IActionResult> DeletePhoto(int userId, int id)
+        {
+            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
+                return Unauthorized();
+
+            var user = await _repo.GetUser(userId);
+
+            if (!user.Photos.Any(p => p.Id == id))
+                return Unauthorized();
+
+            var photoFromRepo = await _repo.GetPhoto(id);
+
+            if (photoFromRepo.IsMain)
+                return BadRequest("You cannot delete your main photo");
+
+            if (photoFromRepo.PublicId != null)
+            {
+                var deleteParams = new DeletionParams(photoFromRepo.PublicId);
+
+                var result = _cloudinary.Destroy(deleteParams);
+
+                if (result.Result == "ok")
+                {
+                    _repo.Delete(photoFromRepo);
+                }
+            }
+
+            if (photoFromRepo.PublicId == null)
+            {
+                _repo.Delete(photoFromRepo);
+            }
+
+            if (await _repo.SaveAll())
+                return Ok();
+
+            return BadRequest("Failed to delete the photo");
+        }
+
     }
+
+    
 }
