@@ -1,21 +1,21 @@
 import { Injectable } from '@angular/core';
-import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, Router } from '@angular/router';
-import { Observable } from 'rxjs/Observable';
+import { CanActivate, Router } from '@angular/router';
 import { AuthService } from '../_services/auth.service';
-import { AlertConfig } from 'ngx-bootstrap/alert/alert.config';
 import { AlertifyService } from '../_services/alertify.service';
 
-@Injectable()
+@Injectable({
+  providedIn: 'root'
+})
 export class AuthGuard implements CanActivate {
+  constructor(private authService: AuthService, private router: Router,
+    private alertify: AlertifyService) {}
 
-  constructor(private authService: AuthService, private router: Router, private alertify: AlertifyService) {}
-
-  canActivate(): Observable<boolean> | Promise<boolean> | boolean {
+  canActivate(): boolean {
     if (this.authService.loggedIn()) {
       return true;
     }
 
-    this.alertify.error('You need to be logged in to access this area');
+    this.alertify.error('You shall not pass!!!');
     this.router.navigate(['/home']);
     return false;
   }
