@@ -1,15 +1,10 @@
-
-
-import { environment } from './../../../environments/environment';
 import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { Photo } from '../../_models/photo';
-
-
+import { FileUploader } from 'ng2-file-upload';
+import { environment } from '../../../environments/environment';
 import { AuthService } from '../../_services/auth.service';
 import { UserService } from '../../_services/user.service';
 import { AlertifyService } from '../../_services/alertify.service';
-import { FileUploader } from 'ng2-file-upload';
-
 
 @Component({
   selector: 'app-photo-editor',
@@ -17,72 +12,71 @@ import { FileUploader } from 'ng2-file-upload';
   styleUrls: ['./photo-editor.component.css']
 })
 export class PhotoEditorComponent implements OnInit {
-  @Input()photos: Photo[];
-  @Output() getMainPhotoUrl = new EventEmitter<string>();
+  @Input() photos: Photo[];
+  @Output() getMemberPhotoChange = new EventEmitter<string>();
   uploader: FileUploader;
   hasBaseDropZoneOver = false;
-  hasAnotherDropZoneOver = false;
-  private baseUrl = environment.apiUrl;
-  private currentMainPhoto: Photo;
+  baseUrl = environment.apiUrl;
+  currentMain: Photo;
 
-  constructor(
-    private authService: AuthService,
-    private _us: UserService,
-    private _as: AlertifyService
-    ) { }
+  constructor(private authService: AuthService, private userService: UserService,
+    private alertify: AlertifyService) { }
 
   ngOnInit() {
-    this.iniciarUploader();
+    this.initializeUploader();
   }
 
-
-  public fileOverBase(e: any): void {
+  fileOverBase(e: any): void {
     this.hasBaseDropZoneOver = e;
   }
 
-  iniciarUploader () {
-    this.uploader = new FileUploader(
-      {
-        allowedFileType : ['image'],
-        isHTML5 : true,
-        maxFileSize: 10 * 1024 * 1024,
-        removeAfterUpload : true,
-        autoUpload: false,
-        url: this.baseUrl + 'users/' + this.authService.decodedToken.nameid + '/photos',
-        authToken : 'Bearer ' + localStorage.getItem('token')
-      }
-    );
-    this.uploader.onAfterAddingFile = (file) => file.withCredentials = false;
+  initializeUploader() {
+    this.uploader = new FileUploader({
+      url: this.baseUrl + 'users/' + this.authService.decodedToken.nameid + '/photos',
+      authToken: 'Bearer ' + localStorage.getItem('token'),
+      isHTML5: true,
+      allowedFileType: ['image'],
+      removeAfterUpload: true,
+      autoUpload: false,
+      maxFileSize: 10 * 1024 * 1024
+    });
 
-    this.uploader.onSuccessItem = (item, response, status) => {
-      const resp: Photo = JSON.parse(response);
-      if (resp) {
+    this.uploader.onSuccessItem = (item, response, status, headers) => {
+      if (response) {
+        const res: Photo = JSON.parse(response);
         const photo = {
-          id : resp.id,
-          url: resp.url,
-          isMain: resp.isMain,
-          dateAdded: resp.dateAdded,
-          description: resp.description
+          id: res.id,
+          url: res.url,
+          dateAdded: res.dateAdded,
+          description: res.description,
+          isMain: res.isMain
         };
-
         this.photos.push(photo);
       }
     };
-
   }
 
-  updateMainPhotoUser(photo: Photo) {
+  setMainPhoto(photo: Photo) {
+    this.userService.setMainPhoto(this.authService.decodedToken.nameid, photo.id).subscribe(() => {
+      this.currentMain = this.photos.filter(p => p.isMain === true)[0];
+      this.currentMain.isMain = false;
+      photo.isMain = true;
+      this.authService.changeMemberPhoto(photo.url);
+      this.authService.currentUser.photoUrl = photo.url;
+      localStorage.setItem('user', JSON.stringify(this.authService.currentUser));
+    }, error => {
+      this.alertify.error(error);
+    });
+  }
 
-    this._us.setMainPhotoUser(this.authService.decodedToken.nameid, photo.id).subscribe(
-      res => {
-        this.currentMainPhoto = this.photos.filter(p => p.isMain === true)[0];
-        this.currentMainPhoto.isMain = false;
-        photo.isMain = true;
-        this.getMainPhotoUrl.emit(photo.url);
-      },
-      err => {
-        this._as.error(err);
-      }
-    );
+  deletePhoto(id: number) {
+    this.alertify.confirm('Are you sure you want to delete this photo?', () => {
+      this.userService.deletePhoto(this.authService.decodedToken.nameid, id).subscribe(() => {
+        this.photos.splice(this.photos.findIndex(p => p.id === id), 1);
+        this.alertify.success('Photo has been deleted');
+      }, error => {
+        this.alertify.error('Failed to delete the photo');
+      });
+    });
   }
 }
