@@ -10,31 +10,37 @@ import { Router } from '@angular/router';
 })
 export class NavComponent implements OnInit {
   model: any = {};
+  photoUrl: string;
 
-  constructor(private authService: AuthService, private alertify: AlertifyService, private router: Router) { }
+  constructor(public authService: AuthService, private alertify: AlertifyService,
+      private router: Router) { }
 
   ngOnInit() {
+    this.authService.currentPhotoUrl.subscribe(photoUrl => this.photoUrl = photoUrl);
   }
 
   login() {
-    this.authService.login(this.model).subscribe(data => {
-      this.alertify.success('logged in successfully');
+    this.authService.login(this.model).subscribe(next => {
+      this.alertify.success('Logged in successfully');
     }, error => {
-      this.alertify.error('Failed to log in');
+      this.alertify.error(error);
     }, () => {
       this.router.navigate(['/members']);
     });
   }
 
-  logout() {
-    this.authService.userToken = null;
-    localStorage.removeItem('token');
-    this.alertify.message('logged out');
-    this.router.navigate(['/home']);
+  loggedIn() {
+    const token = localStorage.getItem('token');
+    return !!token;
   }
 
-  loggedIn() {
-    return this.authService.loggedIn();
+  logout() {
+    localStorage.removeItem('token');
+    localStorage.removeItem('user');
+    this.authService.decodedToken = null;
+    this.authService.currentUser = null;
+    this.alertify.message('logged out');
+    this.router.navigate(['/home']);
   }
 
 }
