@@ -1,4 +1,4 @@
-import { Routes } from '@angular/router';
+import {Routes} from '@angular/router';
 import { HomeComponent } from './home/home.component';
 import { MemberListComponent } from './members/member-list/member-list.component';
 import { MessagesComponent } from './messages/messages.component';
@@ -12,19 +12,21 @@ import { MemberEditResolver } from './_resolvers/member-edit.resolver';
 import { PreventUnsavedChanges } from './_guards/prevent-unsaved-changes.guard';
 
 export const appRoutes: Routes = [
-    { path: 'home', component: HomeComponent},
+    {path: 'home', component: HomeComponent},
     {
         path: '',
         runGuardsAndResolvers: 'always',
         canActivate: [AuthGuard],
         children: [
-            { path: 'members', component: MemberListComponent, resolve: {users: MemberListResolver} },
-            { path: 'members/:id', component: MemberDetailComponent, resolve: {user: MemberDetailResolver} },
-            { path: 'member/edit', component: MemberEditComponent, 
-                resolve: {user: MemberEditResolver}, canDeactivate: [PreventUnsavedChanges] },
-            { path: 'messages', component: MessagesComponent},
-            { path: 'lists', component: ListsComponent},
+            {path: 'members', component: MemberListComponent,
+                resolve: {users: MemberListResolver}},
+            {path: 'members/:id', component: MemberDetailComponent,
+                resolve: {user: MemberDetailResolver}},
+            {path: 'member/edit', component: MemberEditComponent,
+                resolve: {user: MemberEditResolver}, canDeactivate: [PreventUnsavedChanges]},
+            {path: 'messages', component: MessagesComponent},
+            {path: 'lists', component: ListsComponent},
         ]
     },
-    { path: '**', redirectTo: 'home', pathMatch: 'full'},
+    {path: '**', redirectTo: 'home', pathMatch: 'full'},
 ];
